Adding clustering with k=50
Towards confidence_score

Clustering the Dmetaphone hashes with 50 most metaphones as centroids.
Each cluster is persisted in Redis, ranked by their edit distance. The
idea is to use the cluster to classify the Dmetaphone of a new word into
one of the cluster. We can potentially derive a confidence score as to
whether the incoming word is potentially a new word or is a spelling
error.

TODO: Write/Document the Math, as to why this works